add metrics gauge for pending request count
Allows for monitoring of the number of pending operations.

Changes to support this are more extensive than might be expected since
unlike other metrics that can be "pushed" into the Metrics
implementation, there needs to be a way for the Metrics implementation
to "pull" the pending operation count when it is interested in reporting
the values (otherwise the DefaultRawMemcacheClient would have to inform
the Metrics implementation every single time the pending operation
counter changed value).