add folsom-semantic-metrics module
This is a port of internal code we developed at Spotify that implements
Folsom's `Metrics` interface with [semantic-metrics][].

The module was written before semantic-metrics was open sourced, so it
was not added to folsom at the time of writing but kept internal.

There isn't any real reason to keep it internal any longer so this
change adds it to folsom so that things like versions or API changes can
be managed by having this module in the same source tree as folsom
itself.

[semantic-metrics]: https://github.com/spotify/semantic-metrics