Support configurable batch size when creating clients
added `withRequestBatchSize(int)` method to MemcacheClientBuilder.

this allows fine-tuning performance for high-throughput services.
we have been using this fix in highly loaded services in prod env
for years (our local clone of this library).

extracted default batch size constant from BatchFlusher class
to Settings because it was not visible to code outside of that package
and it was required in MemcacheClientBuilder