Add OpenCensus tracing
This add support for tracing memcache operations, including an
implementation for OpenCensus. This change opts to only trace high level
operations for the sake of simplicity. It will for example not trace
operations to individual nodes (e.g. for delete all operations). As we
learn more about tracing needs, we might change this to include more
fine-grained operations.

The OpenCensus implementation attempts to use the OpenTracing
conventions for attributes:
https://opentracing.io/specification/conventions/